+++ docs/CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Breaking
 
-* Change `FlexibleComposeOrder` can also change jongseong order [#534](https://github.com/Riey/kime/issues/534)
+* `FlexibleComposeOrder` can change jongseong order [#534](https://github.com/Riey/kime/issues/534)
 * Implement hanja candidate window [#383](https://github.com/Riey/kime/issues/383)
+* Builtin sebeolsik `3-90` and `3-91` changed with all jungseong to uncomposable except for `ㅑ`, `ㅕ`,`ㅡ`, `ㅜ(9)`, `ㅗ(/)` [#542](https://github.com/Riey/kime/issues/542)
 
 ### Improve
 
+* `FlexibleComposeOrder` can change compose jungseong order [#542](https://github.com/Riey/kime/issues/542)
 * Fix preedit character error on chromium family [#535](https://github.com/Riey/kime/issues/535)
 * Let incomplete character can commit multiple jamos
 * Support johab encoding for preedit string